requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
setuptools-rust
